Install Dotfiles (#7)
# Pull Request

## Description

This pull request includes several changes to improve the setup and
configuration of the development environment, particularly focusing on
the integration of `Justfile` and the setup of dotfiles. The most
important changes include modifications to the
`.github/workflows/code-checks.yml` file, the addition of new commands
in the `Justfile`, and updates to the `development-environment/.bashrc`.

### Workflow and setup improvements:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L67-R67):
Modified the command for checking the Justfile format to use `sh` as the
shell.

### Justfile enhancements:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR1-R10):
Added a shell configuration for `powershell.exe` and included a new
command to install dotfiles using `xcopy`.

### Environment configuration:

*
[`development-environment/.bashrc`](diffhunk://#diff-720dffb230eddc170da91e87acc8112d914c37ff95efdb17898c92dc7c28c238R1-R4):
Added a script to initialize `oh-my-posh` and update the `PATH`
environment variable.