Add Development Scripts (#11)
# Pull Request

## Description

This pull request includes various improvements and additions to the
development environment, focusing on enhancing the `.bashrc`
configuration and adding several useful scripts for Git operations.

### Enhancements to `.bashrc`:

* Changed the path to `oh-my-bash` installation to use the `USER_PATH`
variable instead of a hardcoded path.
* Added multiple aliases for common development tasks, such as reloading
the `.bashrc`, running Prettier, and executing various Git operations.

### Addition of Git operation scripts:

* Added `checkout_main_and_pull.sh` script to reset changes, switch to
the main branch, and pull updates from the remote repository.
* Added `commit_and_push.sh` script to commit staged changes with a
provided commit message and push them to the remote repository.
* Added `rebase_and_push.sh` script to fetch the latest changes, rebase
the current branch with a specified branch, and push the changes with
force-with-lease.
* Added `remove_all_branches_except_default.sh` script to delete all
local branches except for `main` and `master`.
* Added `update_and_push.sh` script to commit changes with a default
message "update" and push them to the remote repository.

### Other improvements:

* Modified `.bash_profile` to use a more robust way of sourcing
`.bashrc` if it exists.