Add EditorConfig Check Git Hook (#96)
# Pull Request

## Description

This pull request introduces a new check for EditorConfig compliance in
the project. The changes include adding a new task in the `Justfile` and
integrating it into the pre-commit hooks configuration.

### EditorConfig Integration:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR84-R89):
Added a new `editorconfig-check` task to run the EditorConfig checker.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR25-R26):
Integrated the `editorconfig-check` task into the pre-commit hooks under
"EditorConfig Checks."